update protocol + continue implementation for world server

+++ Core/Rathalos.Core.Protocol/Messages/Csproto/Classes/CSLoadEntityReq.cs
@@ -16,7 +16,7 @@ namespace Rathalos.Core.Protocol.Messages.Csproto
 		/// <summary>
 		/// Protocol identifier for this message type
 		/// </summary>
-		public ushort ProtocolId => (ushort)ProtocolIdConst;
+		public int ProtocolId => (int)ProtocolIdConst;
 		public uint[] LogicEntityID { get; set; } = Array.Empty<uint>();
 		public uint[] LogicEntityType { get; set; } = Array.Empty<uint>();
