+++ Core/Rathalos.Core.Protocol/Messages/Csproto/Classes/CSSelectRoleRsp.cs
@@ -16,7 +16,7 @@ namespace Rathalos.Core.Protocol.Messages.Csproto
 		/// <summary>
 		/// Protocol identifier for this message type
 		/// </summary>
-		public ushort ProtocolId => (ushort)ProtocolIdConst;
+		public int ProtocolId => (int)ProtocolIdConst;
 		/// <summary>
 		/// 0 for success, -3 means no available TownServer, otherwise error code
 		/// </summary>
