+++ Core/Rathalos.Core.Protocol/Messages/TqqApi/Classes/TPDUExtChgSkey.cs
@@ -16,7 +16,7 @@ namespace Rathalos.Core.Protocol.Messages.Tqqapi
 		/// <summary>
 		/// Protocol identifier for this message type
 		/// </summary>
-		public ushort ProtocolId => (ushort)ProtocolIdConst;
+		public int ProtocolId => (int)ProtocolIdConst;
 		/// <summary>
 		/// Default fill zero
 		/// </summary>
